Rename AnalyticsMock -> AnalyticsMobile
